Make this component work as a climate platform
I didn't understand initially how this was meant to be done. It makes
sense in hindsight.
